requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==9.5.0 